+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.7 branch
 ### 1.7.4
-* all methods provided by the `Filesystem` class (except for `makePathRelative()`) can now be called statically.
+* all methods provided by the `Filesystem` class (except for `makePathRelative()`) can now be called statically;
+* `objects_map()` global function is now deprecated and will be removed in a later release.
 
 ### 1.7.3
 * fixed a bug for `Exceptionist::isInstanceOf()`: this method now accepts instantiated objects or class name as string